refactor: move normalizeAnalysisResponse to entities/analysis layer + test assertion
Move the pure (React-free) normalizeAnalysisResponse transform from
widgets/analysis/utils to the entities/analysis slice where it belongs
per FSD, and export it via the entities barrel. Update useAnalysis to
import from @/entities/analysis.

Reword the JSDoc to accurately describe defense-in-depth: the normalizer
re-establishes the array/object contract at the data source, while
AnalysisPanel and buildExpertAnalysisReport keep independent defensive
defaults for partial responses received directly.

Fix the integration test 'calls onModelChange when the model selector is
used' (MISTAKES Tests §9): it promised an onModelChange call but never
asserted one. Stop mocking usePopoverToggle so the dropdown actually
opens, click a model option, and assert onModelChange was called once
with the selected model id.
